パスを修正。
