+++ app/Http/Controllers/Api/Instructor/InstructorController.php
@@ -7,8 +7,8 @@ use App\Exceptions\DuplicateAuthorizationTokenException;
 use App\Exceptions\ExpiredAuthorizationCodeException;
 use App\Exceptions\TryCountOverAuthorizationCodeException;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Instructor\Instructor\StoreRequest;
-use App\Http\Requests\Instructor\Instructor\UpdateRequest;
+use App\Http\Requests\Instructor\StoreRequest;
+use App\Http\Requests\Instructor\UpdateRequest;
 use App\Http\Requests\Instructor\UserAuthenticationRequest;
 use App\Http\Resources\Instructor\InstructorShowResource;
 use App\Mail\AuthenticationConfirmationMail;
