+++ app/Http/Requests/Instructor/StoreRequest.php
@@ -1,12 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Instructor\Instructor;
+namespace App\Http\Requests\Instructor;
 
-use App\Rules\InstructorUniqueEmailRule;
 use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Support\Facades\Auth;
 
-class UpdateRequest extends FormRequest
+class StoreRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -29,8 +27,7 @@ class UpdateRequest extends FormRequest
             'nick_name' => ['required', 'string', 'max:50'],
             'last_name' => ['required', 'string', 'max:50'],
             'first_name' => ['required', 'string', 'max:50'],
-            'email' => ['required', 'email', new InstructorUniqueEmailRule(Auth::user()->email), 'max:255'],
-            'profile_image' => ['mimes:jpg,png', 'max:2048'],
+            'email' => ['required', 'email', 'max:255', 'unique:instructors'],
         ];
     }
 }
